+++ apps/web/components/SoundNFT.tsx
@@ -1,45 +1,85 @@
-import React from 'react'
-import { NFTData } from '../interfaces/interfaces'
-import { 
-    Flex,
-    SimpleGrid,
-    Box, 
-    Stack,
-    Heading,
-    Text,
-    Spacer
-} from '@chakra-ui/react'
-import dynamic from 'next/dynamic';
+import { Flex, Heading, Spacer, Stack, Text } from "@chakra-ui/react";
+import dynamic from "next/dynamic";
+import React from "react";
+import { NFTData } from "../interfaces/interfaces";
 
-const PlayerGlitch = dynamic(() => import("../components/playerGlitch"),
-  { ssr: false }
-);
 
-export default function SoundNFT(soundNFTData: { soundNFTData: NFTData } ) {
+export default function SoundNFT({nftData}: {nftData?: NFTData})
+{
+    let Player: React.ComponentType<{
+        size: {
+            width: number;
+            height: number;
+        };
+        isDark: boolean;
+    }>
 
-    return (
-        <Flex
-            direction={["column", "column", "row"]}
-            align="center"
-            minH="80vh"
-            minW="80vw"
-        >
-           
-                <Box mt={["15vh", "15vh", "unset"]}>
-                    <PlayerGlitch size={{width:450, height:450}} isDark={false}/>
-                </Box>
-                <Spacer></Spacer>
-                <Stack w={["85vw", "85vw", "30vw"]}>
-                    <Flex display="column">
-                        <Heading size="md">name:</Heading><Text>DEVCSNFT0</Text>
-                        <Heading size="md">probability:</Heading><Text>0.00000 %</Text>
-                        <Heading size="md">policy:</Heading><Text>be3a4e111a307643783820c2bf15fcace87f161187be9301857b593a</Text>
-                        <Heading size="md">sounds:</Heading><Text>sound1, sound2, sound3, sound4, sound5</Text>
-                        <Heading size="md">player:</Heading><Text>glitch</Text>
-                        <Heading size="md">buying xx:</Heading><Text>randomTxHash000000111122222333344445555666677777888889999</Text>
-                        <Heading size="md">mint tx:</Heading><Text>randomTxHash000000111122222333344445555666677777888889999</Text>
+
+    switch(nftData?.metadata?.player)
+    {
+        case "glitch":
+            {
+                Player = dynamic(() => import("./playerGlitch"),
+                                    { ssr: false }
+                                );
+                break
+            }
+            case "superformula":
+            {
+                Player = dynamic(() => import("./playerSuperFormula"),
+                                    { ssr: false }
+                                );
+                break
+            }
+            case "randomDancers":
+            {
+                Player = dynamic(() => import("./playerRandomDancers"),
+                                    { ssr: false }
+                                );
+                break
+            }
+            case "colors":
+            {
+                Player = dynamic(() => import("./playerColors"),
+                                    { ssr: false }
+                                );
+                break
+            }
+        default:
+            Player = null
+            break
+        }
+        
+        return (
+            <>
+            
+                {nftData ?
+                <Flex minH="60vh">
+                    <Player size={{height: 400, width: 400}} isDark={false}/>
+                    <Spacer></Spacer>
+                    <Stack 
+                        minW={["85vw", "85vw", "30vw"]}
+                        align="center"
+                        justify="center"
+                    >
+                        <Flex display="column">
+                            <Heading size="sm">token name:</Heading><a><Text>{nftData.metadata.token_name}</Text></a>
+                            <Heading size="sm">policy:</Heading><Text>be3a4e111a307643783820c2bf15fcace87f161187be9301857b593a</Text>
+                            <Heading size="sm">web:</Heading><a><Text>{nftData.metadata.arweave_website_uri}</Text></a>
+                            <Heading size="sm">rarity color:</Heading><Text>{nftData.metadata.rarity}</Text>
+                            <Heading size="sm">probability:</Heading><Text>{nftData.metadata.probability} %</Text>
+                            <Heading size="sm">sounds:</Heading><Text>{nftData.metadata.sounds.map(x => <p>x.filename</p>)}</Text>
+                            <Heading size="sm">player:</Heading><Text>{nftData.metadata.player}</Text>
+                            <Heading size="sm">buying tx:</Heading><Text>{nftData.tx_Hash}</Text>
+                        </Flex>
+                    </Stack>
+                </Flex>
+                :
+                <>
+                    <Flex minH="60vh" align="center" justify="center">
+                        <Heading size="md">Couldn't find sound you were looking for...</Heading>
                     </Flex>
-                </Stack>
-        </Flex>
-    )
+                </>}
+		</>
+	)
 }
