+++ apps/web/pages/sound/[tokenname].tsx
@@ -1,103 +1,59 @@
-import { Flex, Heading, Spacer, Stack, Text } from "@chakra-ui/react";
-import dynamic from "next/dynamic";
 import Head from "next/head";
 import React from "react";
 import Layout from "../../components/layout";
+import SoundNFT from "../../components/SoundNFT";
 import { NFTData } from "../../interfaces/interfaces";
-
 
 
 export default function Sound({nftData}: {nftData?: NFTData})
 {
-    let Player: React.ComponentType<{
-        size: {
-            width: number;
-            height: number;
-        };
-        isDark: boolean;
-    }>
-
-
-    switch(nftData?.metadata?.player)
-    {
-        case "glitch":
-            {
-                Player = dynamic(() => import("../../components/playerGlitch"),
-                                    { ssr: false }
-                                );
-                break
-            }
-            case "superformula":
-            {
-                Player = dynamic(() => import("../../components/playerSuperFormula"),
-                                    { ssr: false }
-                                );
-                break
-            }
-            case "randomDancers":
-            {
-                Player = dynamic(() => import("../../components/playerRandomDancers"),
-                                    { ssr: false }
-                                );
-                break
-            }
-            case "colors":
-            {
-                Player = dynamic(() => import("../../components/playerColors"),
-                                    { ssr: false }
-                                );
-                break
-            }
-        default:
-            Player = null
-            break
-        }
-        
+    
         return (
             <>
             <Layout>
                 <Head>
-                    <title>Buy Cardano Sounds NFT</title>
+                    <title>Cardano Sounds NFT - {nftData?.id}</title>
                 </Head>
-                {nftData ?
-                <Flex minH="60vh">
-                    <Player size={{height: 400, width: 400}} isDark={false}/>
-                    <Spacer></Spacer>
-                    <Stack 
-                        minW={["85vw", "85vw", "30vw"]}
-                        align="center"
-                        justify="center"
-                    >
-                        <Flex display="column">
-                            <Heading size="sm">token name:</Heading><a><Text>{nftData.metadata.token_name}</Text></a>
-                            <Heading size="sm">policy:</Heading><Text>be3a4e111a307643783820c2bf15fcace87f161187be9301857b593a</Text>
-                            <Heading size="sm">web:</Heading><a><Text>{nftData.metadata.arweave_website_uri}</Text></a>
-                            <Heading size="sm">rarity color:</Heading><Text>{nftData.metadata.rarity}</Text>
-                            <Heading size="sm">probability:</Heading><Text>{nftData.metadata.probability} %</Text>
-                            <Heading size="sm">sounds:</Heading><Text>{nftData.metadata.sounds.map(x => <p>x.filename</p>)}</Text>
-                            <Heading size="sm">player:</Heading><Text>{nftData.metadata.player}</Text>
-                            <Heading size="sm">buying tx:</Heading><Text>{nftData.tx_Hash}</Text>
-                        </Flex>
-                    </Stack>
-                </Flex>
-                :
-                <>
-                    <Flex minH="60vh" align="center" justify="center">
-                        <Heading size="md">Couldn't find sound you were looking for...</Heading>
-                    </Flex>
-                </>}
+                <SoundNFT nftData={nftData} />
+
             </Layout>
-		</>
-	)
+            </>
+	    )
 }
 
 export const getServerSideProps = async (context) => {
     // ...
     const { tokenname } = context.query
 
-    let data: NFTData
+    let data: NFTData = {
+        amount: [{quantity: 25000000, unit: "lovelace"}],
+        id: "CSNFT1",
+        output_Index: 0,
+        senderAddress: "addr112233334444555666678777788888999911100000",
+        metadata: {
+            arweave_id_sound: "hjdf92o3heohdj293hjo2hij3hj0pihjn09",
+            ipfs_id_sound: "ipfs://",
+            image: "ipfs://",
+            player: "randomdancers-light",
+            id: "CSNFT1",
+            probability: 0.001,
+            rarity: "",
+            sounds: [
+                {category: "enrichment", probability: 0.01, filename: "sound1"},
+                {category: "melody", probability: 0.01, filename: "sound2"},
+                {category: "bass", probability: 0.01, filename: "sound3"},
+                {category: "enrichment", probability: 0.01, filename: "sound4"},
+                {category: "enrichment", probability: 0.01, filename: "sound5"}
+            ],
+            token_name: "CSNFT1",
+            arweave_website_uri: ""        
+        },
+        status: "done",
+        tx_Hash: "poj32ohjdf92o3heohdj293hjo2hij3hj0pihjn09o3hdoihwohj02",
+        created: "8/4/2021 10:00pm"
+    }
 
-    data = await fetch("/api/sound/" + tokenname).then(res => res.json())
+    //data = await fetch("/api/sound/" + tokenname).then(res => res.json())
     
     return {
         props: {nftData: data}
